+++ test/test.js
@@ -1,2 +1,4 @@
 import calc from './calc'
 import utils from './utils'
+import Process from './Process'
+import loop from './loop'